Conflicts:
	public/js/production/application.min.js.map

+++ public/js/add.projects.js
@@ -372,6 +372,7 @@
         return;
       }
       floorSelection = $(this).closest('.row').find('.select-floor select');
+      floorSelection.select2('val', '');
       noOfFloors = $(this).find('option[value="' + buildingId + '"]').attr('data-no-of-floors');
       if (parseInt(noOfFloors) === 0) {
         return;
@@ -593,6 +594,44 @@
     };
     return $.ajax({
       url: "/admin/project/" + PROJECTID + "/building/" + buildingId,
+      type: 'DELETE',
+      success: successFn
+    });
+  });
+
+  $('.delete-varint').click(function() {
+    var successFn, variantId, variantType;
+    if (confirm('Are you sure you want to delete this variant?') === false) {
+      return;
+    }
+    variantId = $(this).attr('data-variant-id');
+    variantType = $(this).attr('data-variant-type');
+    successFn = function(resp, status, xhr) {
+      if (xhr.status === 204) {
+        return window.location = "/admin/project/" + PROJECTID + "/" + variantType + "";
+      }
+    };
+    return $.ajax({
+      url: "/admin/project/" + PROJECTID + "/bunglow-variant/" + variantId,
+      type: 'DELETE',
+      success: successFn
+    });
+  });
+
+  $('.delete-unit').click(function() {
+    var successFn, unitId, unitType;
+    if (confirm('Are you sure you want to delete this unit?') === false) {
+      return;
+    }
+    unitId = $(this).attr('data-unit-id');
+    unitType = $(this).attr('data-unit-type');
+    successFn = function(resp, status, xhr) {
+      if (xhr.status === 204) {
+        return window.location = "/admin/project/" + PROJECTID + "/" + unitType + "";
+      }
+    };
+    return $.ajax({
+      url: "/admin/project/" + PROJECTID + "/bunglow-unit/" + unitId,
       type: 'DELETE',
       success: successFn
     });
